Move hook to selector
