+++ packages/edit-site/src/components/template-details/template-areas.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import { keyBy } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
@@ -11,12 +6,11 @@ import { MenuGroup, MenuItem } from '@wordpress/components';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { getTemplatePartIcon } from '@wordpress/editor';
 import { store as blockEditorStore } from '@wordpress/block-editor';
-import { store as coreDataStore } from '@wordpress/core-data';
-import { isTemplatePart } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
  */
+import { store as editSiteStore } from '../../store';
 import { TEMPLATE_PART_AREA_TO_NAME } from '../../store/constants';
 
 function TemplatePartItem( { area, clientId } ) {
@@ -45,38 +39,8 @@ function TemplatePartItem( { area, clientId } ) {
 
 export default function TemplateAreas( { template } ) {
 	const templateAreaBlocks = useSelect(
-		( select ) => {
-			const templateParts = select( coreDataStore ).getEntityRecords(
-				'postType',
-				'wp_template_part',
-				{
-					per_page: -1,
-				}
-			);
-			const templatePartsById = keyBy(
-				templateParts,
-				( templatePart ) => templatePart.id
-			);
-
-			const templatePartBlocksByAreas = {};
-
-			for ( const block of template.blocks ) {
-				if ( isTemplatePart( block ) ) {
-					const {
-						attributes: { theme, slug },
-					} = block;
-					const templatePartId = `${ theme }//${ slug }`;
-					const templatePart = templatePartsById[ templatePartId ];
-
-					if ( templatePart ) {
-						templatePartBlocksByAreas[ templatePart.area ] = block;
-					}
-				}
-			}
-
-			return templatePartBlocksByAreas;
-		},
-		[ template.blocks ]
+		( select ) => select( editSiteStore ).getTemplateAreaBlocks( template ),
+		[ template ]
 	);
 
 	return (
