+++ packages/edit-site/src/store/selectors.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { get, map } from 'lodash';
+import { get, map, keyBy } from 'lodash';
 import createSelector from 'rememo';
 
 /**
@@ -10,6 +10,7 @@ import createSelector from 'rememo';
 import { store as coreDataStore } from '@wordpress/core-data';
 import { createRegistrySelector } from '@wordpress/data';
 import { uploadMedia } from '@wordpress/media-utils';
+import { isTemplatePart } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
@@ -260,3 +261,44 @@ export function __experimentalGetInsertionPoint( state ) {
 export function isListViewOpened( state ) {
 	return state.listViewPanel;
 }
+
+/**
+ * Returns the template part blocks grouped by areas for a template.
+ *
+ * @param {Object} state    Global application state.
+ * @param {Object} template The template object.
+ * @return {Object} Template part blocks by areas.
+ */
+export const getTemplateAreaBlocks = createRegistrySelector(
+	( select ) => ( state, template ) => {
+		const templateParts = select( coreDataStore ).getEntityRecords(
+			'postType',
+			'wp_template_part',
+			{
+				per_page: -1,
+			}
+		);
+		const templatePartsById = keyBy(
+			templateParts,
+			( templatePart ) => templatePart.id
+		);
+
+		const templatePartBlocksByAreas = {};
+
+		for ( const block of template.blocks ) {
+			if ( isTemplatePart( block ) ) {
+				const {
+					attributes: { theme, slug },
+				} = block;
+				const templatePartId = `${ theme }//${ slug }`;
+				const templatePart = templatePartsById[ templatePartId ];
+
+				if ( templatePart ) {
+					templatePartBlocksByAreas[ templatePart.area ] = block;
+				}
+			}
+		}
+
+		return templatePartBlocksByAreas;
+	}
+);
